requirement.txt
langchain
langchain-community
PyPDF2
python-dotenv
streamlit
faiss-cpu
tiktoken
openai
langchain-openai
python-docx